+++ gameplay-samples/sample04-particles/src/ParticlesGame.h
@@ -13,7 +13,7 @@ class ParticlesGame: public Game, Control::Listener
 public:
 
     /**
-     * Constructror.
+     * Constructor.
      */
     ParticlesGame();
 
@@ -47,12 +47,12 @@ protected:
     /**
      * @see Game::update
      */
-    void update(long elapsedTime);
+    void update(float elapsedTime);
 
     /**
      * @see Game::render
      */
-    void render(long elapsedTime);
+    void render(float elapsedTime);
 
 private:
